xAImicroservices/ANCHORS/ANCHORS-Generate/requirements.txt
attrs==23.1.0
backcall==0.2.0
colorama==0.4.6
decorator==5.1.1
dill==0.3.6
h5py==3.8.0
ipython==8.11.0
jedi==0.18.2
joblib==1.2.0
jsonschema==4.17.3
matplotlib-inline==0.1.6
numpy==1.23.5
numba==0.56.4
pandas==1.5.3
parso==0.8.3
pickleshare==0.7.5
prompt-toolkit==3.0.38
Pygments==2.14.0
pyrsistent==0.19.3
python-dateutil==2.8.2
pytz==2023.3
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
threadpoolctl==3.1.0
tqdm==4.65.0
traitlets==5.9.0
typing_extensions==4.5.0
wcwidth==0.2.6
zipp==3.15.0
requests==2.28.2
requests-oauthlib==1.3.1
requests-toolbelt==0.10.1
keras==2.12.0
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.32.0


pymongo==4.3.3
confluent-kafka==2.1.1
paho-mqtt==1.6.1

alibi==0.9.4